+++ CHANGELOG.md
@@ -5,8 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
-###Added
+### Added
 - Update Group API to hook polls  [#1617](https://github.com/rokwire/illinois-app/issues/1617).
+### Changed
+- Use Gateway BB for loading laundries [#1530](https://github.com/rokwire/illinois-app/issues/1530).
 
 ## [4.0.5] - 2022-05-09
 ### Fixed